--- a/python/nutpie/__init__.py
+++ b/python/nutpie/__init__.py
@@ -1,7 +1,14 @@
 from nutpie import _lib
 from nutpie.compile_pymc import compile_pymc_model
 from nutpie.compile_stan import compile_stan_model
-from nutpie.sample import sample
+from nutpie.compile_onnx import compile_pytensor_module
+from nutpie.sampling import sample
 
 __version__: str = _lib.__version__
-__all__ = ["__version__", "sample", "compile_pymc_model", "compile_stan_model"]
+__all__ = [
+    "__version__",
+    "sample",
+    "compile_pymc_model",
+    "compile_stan_model",
+    "compile_pytensor_module",
+]