--- a/src/lib.rs
+++ b/src/lib.rs
@@ -1,7 +1,13 @@
+#[cfg(feature = "iree")]
+mod iree;
+#[cfg(feature = "onnx")]
+mod ort;
 mod progress;
 mod pyfunc;
 mod pymc;
 mod stan;
+#[cfg(feature = "torch")]
+mod torch;
 mod wrapper;
 
 pub use wrapper::_lib;