--- a/src/lib.rs
+++ b/src/lib.rs
@@ -1,5 +1,3 @@
-#[cfg(feature = "iree")]
-mod iree;
 #[cfg(feature = "onnx")]
 mod ort;
 mod progress;